+++ src/app/app-routes.ts
@@ -1,0 +1,9 @@
+export const appRoutes: AppRoutes = {
+  players: 'players',
+  matches: 'matches'
+};
+
+export interface AppRoutes {
+  players: string;
+  matches: string;
+}