+++ src/app/app-routing.module.ts
@@ -1,7 +1,30 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-const routes: Routes = [];
+import { appRoutes } from './app-routes';
+
+const routes: Routes = [
+  {
+    path: appRoutes.players,
+    title: 'Table Tennis Results Tracker • Players',
+    loadChildren: () => import('./views/players/players.module').then((m) => m.PlayersModule)
+  },
+  {
+    path: appRoutes.matches,
+    title: 'Table Tennis Results Tracker • Matches',
+    loadChildren: () => import('./views/matches/matches.module').then((m) => m.MatchesModule)
+  },
+  {
+    path: '',
+    redirectTo: appRoutes.players,
+    pathMatch: 'full'
+  },
+  {
+    path: '**',
+    title: 'Table Tennis Results Tracker • Page Not Found',
+    loadChildren: () => import('./views/page-not-found/page-not-found.module').then((m) => m.PageNotFoundModule)
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],