+++ src/app/layout/layout.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+
+import { ToolbarModule } from './toolbar/toolbar.module';
+
+@NgModule({
+  imports: [ToolbarModule],
+  exports: [ToolbarModule],
+  declarations: [],
+  providers: []
+})
+export class LayoutModule {}