+++ src/app/views/matches/matches-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+
+import { MatchesComponent } from './matches.component';
+
+@NgModule({
+  imports: [
+    RouterModule.forChild([
+      {
+        path: '',
+        component: MatchesComponent
+      }
+    ])
+  ],
+  exports: [RouterModule]
+})
+export class MatchesRoutingModule {}