+++ src/app/views/players/players.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+
+import { PlayersRoutingModule } from './players-routing.module';
+import { PlayersComponent } from './players.component';
+
+@NgModule({
+  imports: [PlayersRoutingModule],
+  exports: [],
+  declarations: [PlayersComponent],
+  providers: []
+})
+export class PlayersModule {}